Formerly, the schemaLocation included both     http://www.w3.org/2005/XQueryX/xqueryx.xsd and     http://www.w3.org/2007/xquery-update-10/xquery-update-10-xqueryx.xsd
However, the latter schema has
    <xsd:import namespace="http://www.w3.org/2005/XQueryX" schemaLocation="
    http://www.w3.org/2007/xquery-update-10/xquery-update-10-xqueryx-redef.xsd
    "/>
and *that* schema has
    <xsd:redefine schemaLocation="http://www.w3.org/2005/XQueryX/xqueryx.xsd">

I.e., the original document is pulling in both
    http://www.w3.org/2005/XQueryX/xqueryx.xsd
and a redefined version of it.

CMSMcQ says "when a schema document is reachable by two different
include/import/redefine paths, the spec is unclear on the result
and processors are all over the map; there is effectively no
interoperability in such situations."

(In particular, I was getting validation errors from a couple of different
validators because they didn't see the redefined version.)

So, remove
    http://www.w3.org/2005/XQueryX
    http://www.w3.org/2005/XQueryX/xqueryx.xsd
from the schemaLocation.